Moving image on screen
Finished the third part of the tutorial series where it teaches how to move an image at the screen. Had to do some refactoring of the functions, and added other members for the Application class. Updated the Makefile to add the flags for SDL2. Furthermore, after the program is run, it shows a warning "queue 0x14dd0d0 destroyed while proxies still attached: wl_callback@42 still attached wl_surface@40 still attached" don't know if that's okay, probably not, but I sent a message at the blog post asking what can be the issue: https://dev.to/noah11012/using-sdl2-scaling-and-moving-an-image-pj