Controlling the Image
On Noah's tutorial, he used a stick figure as the image to display and control on the screen. I'm using a slug as the image instead of the stick figure, I just named the class as StickFigure so that I didn't get lost in his code. But maybe I should change to just Slug.